# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Astra — Autonomous Agent OS Platform API
  version: 1.0.0
  description: |
    Astra is a planet-scale autonomous agent operating system. This specification
    documents every HTTP endpoint across all Astra microservices.

    ## Architecture Overview

    Astra runs as a set of cooperating microservices:
    - **API Gateway** (port 8080) — public-facing, JWT-protected
    - **Goal Service** (port 8088) — goal lifecycle and planning
    - **Identity** (port 8085) — JWT issuance and validation
    - **Access Control** (port 8086) — OPA-like policy checks and approval workflow
    - **Tool Runtime** (port 8083) — sandboxed tool execution
    - **Worker Manager** (port 8082) — worker registry and heartbeats
    - **Prompt Manager** (port 8084) — prompt template storage
    - **Planner Service** (port 8087) — goal-to-task DAG decomposition
    - **Evaluation Service** (port 8089) — task result evaluation
    - **Cost Tracker** (port 8090) — LLM usage cost aggregation

    ## Real-World Use Case: Building an E-Commerce Website

    This walkthrough shows how a user interacts with Astra to autonomously
    build a full Next.js e-commerce site.

    ### Step 1 — Acquire a JWT token
    ```
    POST http://localhost:8085/tokens
    {"subject": "developer", "scopes": ["admin"], "ttl_seconds": 3600}
    ```

    ### Step 2 — Create an e-commerce builder agent
    ```
    POST http://localhost:8080/agents
    Authorization: Bearer <token>
    {
...
    }
...

...

...

...

...

...
servers:
...
security:
  - BearerAuth: []
paths:
...
  /agents:
    post:
      tags: [Gateway — Agents]
      summary: Create a new agent
      description: |
        Spawns a new agent actor in the kernel. Optionally provide a system
        prompt and structured config for the agent's persona.

        **E-Commerce example**: Create an agent that builds websites.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAgentRequest'
            examples:
              ecommerce-builder:
                summary: E-Commerce Builder
                value:
                  actor_type: ecommerce-builder
                  name: E-Commerce Builder
                  system_prompt: "You are a senior full-stack developer specializing in Next.js 14, TypeScript, and Tailwind CSS. You produce clean, production-ready code."
                  config: '{"model_preference":"code"}'
              python-expert:
                summary: Python Expert
                value:
                  actor_type: python-expert
                  name: Python Expert
                  system_prompt: "You are a senior Python expert. You write clean, idiomatic Python (3.10+), follow PEP 8, use type hints, and prefer the standard library. You produce production-ready code with tests when appropriate."
                  config: '{"model_preference":"code"}'
      responses:
        '200':
          description: Agent created
          content:
            application/json:
              example:
                actor_id: "a3ed7085-1b3d-4818-90fa-b54616a45721"
        '400':
          description: Invalid body (e.g. missing actor_type)
        '401':
          description: Missing or invalid JWT
    get:
      tags: [Gateway — Agents]
      summary: List all agents
      description: |
        Returns all agents (summary fields only). Gateway proxies to agent-service.
        Use GET /agents/{id}/profile for full profile including system_prompt and config.
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of agent summaries
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentListResponse'
              example:
                agents:
                  - id: "a3ed7085-1b3d-4818-90fa-b54616a45721"
                    actor_type: ecommerce-builder
                    name: E-Commerce Builder
                    status: active
                    created_at: "2026-03-10T12:00:00Z"
                    updated_at: "2026-03-10T14:30:00Z"
        '401':
          description: Missing or invalid JWT
  /agents/{id}:
    patch:
      tags: [Gateway — Agents]
      summary: Update agent profile
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/AgentId'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                system_prompt:
                  type: string
                config:
                  type: string
            example:
              system_prompt: "You are a principal architect. You design systems but never write code."
      responses:
        '200':
          description: Updated
          content:
            application/json:
              example:
                status: ok
        '400':
          description: Invalid agent id or body
        '401':
          description: Missing or invalid JWT
        '404':
          description: Agent not found
  /agents/{id}/profile:
    get:
      tags: [Gateway — Agents]
      summary: Get agent profile
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/AgentId'
      responses:
        '200':
          description: Agent profile
          content:
            application/json:
              example:
                id: "a3ed7085-1b3d-4818-90fa-b54616a45721"
                name: E-Commerce Builder
                system_prompt: "You are a senior full-stack developer..."
                config: '{"model_preference":"code"}'
        '400':
          description: Invalid agent id
        '401':
          description: Missing or invalid JWT
        '404':
          description: Agent not found

  # ═══════════════════════════════════════════════════════════════════
  # API GATEWAY — AGENT DOCUMENTS
  # ═══════════════════════════════════════════════════════════════════
  /agents/{id}/documents:
    post:
      tags: [Gateway — Agent Documents]
      summary: Attach a document to an agent
      description: |
        Upload a rule, skill, or context document. Documents persist across
        goals unless a `goal_id` is provided.

        **E-Commerce example**: Attach coding standards as a rule.
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/AgentId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDocumentRequest'
            examples:
              rule:
                summary: Coding rule
                value:
                  doc_type: rule
                  name: typescript-only
                  content: "All code must use TypeScript. No .js files. Use strict mode."
                  priority: 10
              skill:
                summary: React skill
                value:
                  doc_type: skill
                  name: nextjs-patterns
                  content: "Use Next.js 14 App Router. Use server components by default. Client components only when needed for interactivity."
                  priority: 20
              context_doc:
                summary: PRD document
                value:
                  doc_type: context_doc
                  name: ecommerce-prd
                  content: "# E-Commerce PRD\n\n## Pages\n- Home with hero + featured products\n- Product listing with filters\n- Product detail with add-to-cart\n- Cart with quantity controls\n- Checkout with form validation\n- Order confirmation"
                  priority: 50
      responses:
        '201':
          description: Document created
          content:
            application/json:
              example:
                id: "d4f5e6a7-b8c9-4d0e-1f2a-3b4c5d6e7f8a"
    get:
      tags: [Gateway — Agent Documents]
      summary: List agent documents
      description: List all documents attached to an agent. Filter by type or goal.
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/AgentId'
        - name: doc_type
          in: query
          schema:
            type: string
            enum: [rule, skill, context_doc, reference]
        - name: goal_id
          in: query
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Document list
          content:
            application/json:
              example:
                documents:
                  - id: "d4f5e6a7-b8c9-4d0e-1f2a-3b4c5d6e7f8a"
                    agent_id: "a3ed7085-1b3d-4818-90fa-b54616a45721"
                    doc_type: rule
                    name: typescript-only
                    content: "All code must use TypeScript..."
                    priority: 10
                    created_at: "2026-03-10T12:00:00Z"
  /agents/{id}/documents/{docId}:
    delete:
      tags: [Gateway — Agent Documents]
      summary: Delete a document
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/AgentId'
        - name: docId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              example:
                status: ok

  # ═══════════════════════════════════════════════════════════════════
  # API GATEWAY — GOALS
  # ═══════════════════════════════════════════════════════════════════
  /agents/{id}/goals:
    post:
      tags: [Gateway — Goals]
      summary: Submit a goal to an agent (via gateway)
      description: |
        Creates a goal and triggers the agent's internal planner. This routes
        through the kernel's actor message system.
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/AgentId'
      requestBody:
        required: true
        content:
          application/json:
            example:
              goal_text: "Build a Next.js 14 e-commerce website with product catalog, shopping cart, checkout flow, and Tailwind CSS styling"
      responses:
        '200':
          description: Goal submitted to agent

  # ═══════════════════════════════════════════════════════════════════
  # API GATEWAY — TASKS & GRAPHS
  # ═══════════════════════════════════════════════════════════════════
  /tasks/{taskId}:
    get:
      tags: [Gateway — Tasks & Graphs]
      summary: Get task details
      security:
        - BearerAuth: []
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Task details
          content:
            application/json:
              example:
                id: "13806f4b-6a7e-4642-b606-6bea145d8f07"
                graph_id: "f635e11c-51b1-4978-ad16-2f50f8597eb5"
                agent_id: "a3ed7085-1b3d-4818-90fa-b54616a45721"
                type: code_generate
                status: completed
                priority: 100
                retries: 0
        '400':
          description: Invalid task id
        '401':
          description: Missing or invalid JWT
        '404':
          description: Task not found
  /tasks/{taskId}/complete:
    post:
      tags: [Gateway — Tasks & Graphs]
      summary: Manually complete a task
      security:
        - BearerAuth: []
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            example:
              result: "Task manually completed by operator"
      responses:
        '200':
          description: Task completed
  /graphs/{graphId}:
    get:
      tags: [Gateway — Tasks & Graphs]
      summary: Get task dependency graph
      description: Returns all tasks and their dependency relationships for a goal's task graph.
      security:
        - BearerAuth: []
      parameters:
        - name: graphId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Graph with tasks and dependencies
          content:
            application/json:
              example:
                tasks:
                  - id: "task-1-uuid"
                    type: shell_exec
                    status: completed
                    priority: 100
                  - id: "task-2-uuid"
                    type: code_generate
                    status: running
                    priority: 100
                dependencies:
                  - task_id: "task-2-uuid"
                    depends_on: "task-1-uuid"
        '400':
          description: Invalid graph id
        '401':
          description: Missing or invalid JWT
        '404':
          description: Graph not found

  # ═══════════════════════════════════════════════════════════════════
  # API GATEWAY — DASHBOARD
  # ═══════════════════════════════════════════════════════════════════
...
  /api/dashboard/approvals/{id}/approve:
    post:
      tags: [Gateway — Dashboard]
      summary: Approve a pending request (from dashboard)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            example:
              decided_by: dashboard-ui
      responses:
        '200':
          description: Approved
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "JWT from POST /tokens on the Identity service"
  parameters:
    AgentId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Agent UUID
  schemas:
    AgentListResponse:
      type: object
      required: [agents]
      properties:
        agents:
          type: array
          description: List of agent summaries
          items:
            $ref: '#/components/schemas/AgentSummary'
...